Add google-java-format to the build script
We use the https://github.com/sherter/google-java-format-gradle-plugin google-java-format plugin from sherter
to ensure formatting in the code and will check it also during CI with this command.

Downside: google-java-format is compiled for Java 11 and cannot be run with jConstraints default java 8 build system.
Code-Formatting is done with a different Java version for now. Lifting JPF to Java 11 should
fix this problem. As this in on the road map, it does not seems to be a real problem.